+++ packages/core/src/storage/StorageContext.ts
@@ -1,5 +1,10 @@
+import * as path from "path";
 import { GenericFileSystem } from "./FileSystem";
-import { DEFAULT_FS, DEFAULT_NAMESPACE } from "./constants";
+import {
+  DEFAULT_FS,
+  DEFAULT_IMAGE_VECTOR_NAMESPACE,
+  DEFAULT_NAMESPACE,
+} from "./constants";
 import { SimpleDocumentStore } from "./docStore/SimpleDocumentStore";
 import { BaseDocumentStore } from "./docStore/types";
 import { SimpleIndexStore } from "./indexStore/SimpleIndexStore";
@@ -11,12 +16,15 @@ export interface StorageContext {
   docStore: BaseDocumentStore;
   indexStore: BaseIndexStore;
   vectorStore: VectorStore;
+  imageVectorStore?: VectorStore;
 }
 
 type BuilderParams = {
   docStore?: BaseDocumentStore;
   indexStore?: BaseIndexStore;
   vectorStore?: VectorStore;
+  imageVectorStore?: VectorStore;
+  storeImages?: boolean;
   persistDir?: string;
   fs?: GenericFileSystem;
 };
@@ -25,6 +33,8 @@ export async function storageContextFromDefaults({
   docStore,
   indexStore,
   vectorStore,
+  imageVectorStore,
+  storeImages,
   persistDir,
   fs,
 }: BuilderParams): Promise<StorageContext> {
@@ -32,6 +42,7 @@ export async function storageContextFromDefaults({
     docStore = docStore || new SimpleDocumentStore();
     indexStore = indexStore || new SimpleIndexStore();
     vectorStore = vectorStore || new SimpleVectorStore();
+    imageVectorStore = storeImages ? new SimpleVectorStore() : imageVectorStore;
   } else {
     fs = fs || DEFAULT_FS;
     docStore =
@@ -45,11 +56,18 @@ export async function storageContextFromDefaults({
       indexStore || (await SimpleIndexStore.fromPersistDir(persistDir, fs));
     vectorStore =
       vectorStore || (await SimpleVectorStore.fromPersistDir(persistDir, fs));
+    imageVectorStore = storeImages
+      ? await SimpleVectorStore.fromPersistDir(
+          path.join(persistDir, DEFAULT_IMAGE_VECTOR_NAMESPACE),
+          fs,
+        )
+      : imageVectorStore;
   }
 
   return {
     docStore,
     indexStore,
     vectorStore,
+    imageVectorStore,
   };
 }